#1226 | Do not reuse live object when editing
Objects being passed in to the form wizard should never be a live object.
One situation where this is not happening is when an enrolment is saved.
While I could not accurately recreate the reported issue, this change
will ensure that the program enrolment object is always a clone when it is
being edited.